+++ main.lua
@@ -7,6 +7,9 @@
 -- hide the status bar
 display.setStatusBar( display.HiddenStatusBar )
 
+-- turn on/off mui debug output
+_mui_debug = true
+
 -- include the Corona "composer" module
 local composer = require "composer"