+++ materialui/mui-button.lua
@@ -364,7 +364,7 @@ function M.touchRRectButton (event)
 
         if M.isTouchPointOutOfRange( event ) then
               event.phase = "offTarget"
-              -- print("Its out of the button area")
+              -- M.debug("Its out of the button area")
               -- event.target:dispatchEvent(event)
         else
             event.phase = "onTarget"
@@ -393,6 +393,7 @@ function M.touchRRectButton (event)
         muiData.interceptMoved = false
         muiData.touching = false
     end
+    return true
 end
 
 --[[
@@ -808,7 +809,7 @@ function M.touchIconButton (event)
         if M.isTouchPointOutOfRange( event ) then
             event.phase = "offTarget"
             -- event.target:dispatchEvent(event)
-            -- print("Its out of the button area")
+            -- M.debug("Its out of the button area")
         else
           event.phase = "onTarget"
             if muiData.interceptMoved == false then
@@ -832,6 +833,7 @@ function M.touchIconButton (event)
             M.deactivateImageTouch( options )
         end
     end
+    return true
 end
 
 function M.createCheckBox(options)
@@ -1048,7 +1050,7 @@ function M.touchCircleButton (event)
         if M.isTouchPointOutOfRange( event ) then
             event.phase = "offTarget"
             -- event.target:dispatchEvent(event)
-            -- print("Its out of the button area")
+            -- M.debug("Its out of the button area")
         else
           event.phase = "onTarget"
             if muiData.interceptMoved == false then
@@ -1071,6 +1073,7 @@ function M.touchCircleButton (event)
             M.deactivateImageTouch( options )
         end
     end
+    return true
 end
 
 --[[
@@ -1326,7 +1329,7 @@ function M.touchCheckbox (event)
         if M.isTouchPointOutOfRange( event ) then
             event.phase = "offTarget"
             -- event.target:dispatchEvent(event)
-            -- print("Its out of the button area")
+            -- M.debug("Its out of the button area")
         else
           event.phase = "onTarget"
             if muiData.interceptMoved == false then
@@ -1350,6 +1353,7 @@ function M.touchCheckbox (event)
             muiData.touching = false
         end
     end
+    return true
 end
 
 function M.createRadioGroup(options)
@@ -1421,7 +1425,7 @@ function M.actionForPlus( e )
             muiTarget.isChecked = true
             muiTarget.text = M.getMaterialFontCodePointByName("add_circle")
             if muiTargetValue ~= nil then
-                print("checkbox value = "..muiTargetValue)
+                M.debug("checkbox value = "..muiTargetValue)
             end
         end
     end
@@ -1439,7 +1443,7 @@ function M.actionForCheckbox( e )
             muiTarget.isChecked = true
             muiTarget.text = M.getMaterialFontCodePointByName("check_box")
             if muiTargetValue ~= nil then
-                print("checkbox value = "..muiTargetValue)
+                M.debug("checkbox value = "..muiTargetValue)
             end
         end
     end
@@ -1469,13 +1473,13 @@ function M.actionForRadioButton( e )
         end
         if muiTargetValue ~= nil then
             muiData.widgetDict[basename]["value"] = muiTargetValue
-            print("radio button value = "..muiTargetValue)
+            M.debug("radio button value = "..muiTargetValue)
         end
     end
 end
 
 function M.actionForButton( e )
-    print("button action!")
+    M.debug("button action!")
 end
 
 function M.removeWidgetRRectButton(widgetName)