+++ materialui/mui-snackbar.lua
@@ -251,7 +251,7 @@ function M.newSnackBar( options )
             if muiData.widgetDict[options.name]["removeInProgress"] == true then return end
             if M.isTouchPointOutOfRange( event ) then
                   event.phase = "offTarget"
-                  -- print("Its out of the button area")
+                  -- M.debug("Its out of the button area")
                   -- event.target:dispatchEvent(event)
             else
                 event.phase = "onTarget"
@@ -269,6 +269,7 @@ function M.newSnackBar( options )
             muiData.interceptMoved = false
             muiData.touching = false
         end
+        return true
     end
     muiData.widgetDict[options.name]["rrect"]:addEventListener( "touch", muiData.widgetDict[options.name]["rrect"] )